Fixes website deploy to gh-pages
Previously part of the deploy process was to run `server.sh` script, which build the project
and run http server with `--watch` option. To continue the deploy one had to press `Ctrl + C`
combination, which was scary and unpleasent for me. I've figured out, that there is a `build`
command for `brunch` which is exactly what builds project when you call `server`  command.
I've compared folders generated with `build` and `server` commands with `diff` and they're
the same, so the script seems pretty safe.